day 5

+++ app/Http/Controllers/ArticleController.php
@@ -4,9 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'detail');
+    }
+
     public function index()
     {
         // $data = Article::all();
@@ -31,14 +38,21 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
+        
+        if(Gate::allows('delete-article', $article)) {
+            $article->delete();
+            return redirect("/articles")->with("info", "Article deleted");
+        }
 
-        return redirect("/articles")->with("info", "Article deleted");
+        return back()->with("info", "Unauthorize");
     }
 
     public function add()
     {
-        return view('articles.add');
+        $category = Category::all();
+        return view('articles.add', [
+            'category' => $category
+        ]);
     }
 
     public function create()
@@ -53,12 +67,23 @@ class ArticleController extends Controller
             return back()->withErrors($validator);
         }
 
+
+
         $article = new Article;
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->id();
         $article->save();
 
         return redirect("/articles");
     }
+
+    public function edit()
+    {
+        
+    }
+
+
+    
 }
